+++ templates/parts/meta/meta-invoice-dates.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.6.0
  */
 
 
@@ -17,38 +17,38 @@
 
 	global $invoice_helper;
 
-	$due_period   = absint( get_theme_mod( 'invoice_due_period', 21 ) );
-	$due_date     = strtotime( '+' . $due_period . ' days', strtotime( $invoice_helper['publish_date_raw'] ) );
-	$due_date     = date( get_option( 'date_format' ), $due_date );
+	$date_format         = get_option( 'date_format' );
+	$date_format_display = Invoices_Helper::get_readable_date_format( $date_format );
 
-	$publish_date = str_replace( '/', '<span class="sep">/</span>', esc_html( $invoice_helper['publish_date_display'] ) );
-	$due_date     = str_replace( '/', '<span class="sep">/</span>', esc_html( $due_date ) );
+	$due_period = absint( $invoice_helper['due_period'] );
+	$due_date   = strtotime( '+' . $due_period . ' days', strtotime( $invoice_helper['publish_date_raw'] ) );
+	$due_date   = date( $date_format, $due_date );
 
 ?>
 
 <div class="invoice-meta-date-container invoice-meta-container">
 
 	<div class="invoice-meta-date invoice-meta-item">
-		<h2 class="invoice-meta-label"><?php esc_html_e( 'Date of issue (d/m/y)', '_invoices' ); ?></h2>
+		<h2 class="invoice-meta-label"><?php printf( esc_html_x( 'Date of issue (%s)', '%s: date format', '_invoices' ), $date_format_display ); ?></h2>
 		<p class="invoice-meta-value">
 			<a href="<?php echo esc_url( get_day_link(
 				date( 'Y', strtotime( $invoice_helper['publish_date_raw'] ) ),
 				date( 'm', strtotime( $invoice_helper['publish_date_raw'] ) ),
 				date( 'd', strtotime( $invoice_helper['publish_date_raw'] ) )
 			) ); ?>">
-				<?php echo $publish_date; ?>
+				<?php echo Invoices_Helper::get_output_date( esc_html( $invoice_helper['publish_date_display'] ) ); ?>
 			</a>
 		</p>
 	</div>
 
 	<div class="invoice-meta-date invoice-meta-item">
-		<h2 class="invoice-meta-label"><?php esc_html_e( 'Date of delivery (d/m/y)', '_invoices' ); ?></h2>
-		<p class="invoice-meta-value"><?php echo $publish_date; ?></p>
+		<h2 class="invoice-meta-label"><?php printf( esc_html_x( 'Date of delivery (%s)', '%s: date format', '_invoices' ), $date_format_display ); ?></h2>
+		<p class="invoice-meta-value"><?php echo Invoices_Helper::get_output_date( esc_html( $invoice_helper['publish_date_display'] ) ); ?></p>
 	</div>
 
-	<div class="invoice-meta-date invoice-meta-item">
-		<h2 class="invoice-meta-label"><?php esc_html_e( 'Due date (d/m/y)', '_invoices' ); ?></h2>
-		<p class="invoice-meta-value"><?php echo $due_date; ?></p>
+	<div class="invoice-meta-date invoice-meta-item" title="<?php printf( esc_html( 'Due period: %d days', '_invoices' ), $due_period ); ?>">
+		<h2 class="invoice-meta-label"><?php printf( esc_html_x( 'Due date (%s)', '%s: date format', '_invoices' ), $date_format_display ); ?></h2>
+		<p class="invoice-meta-value"><?php echo Invoices_Helper::get_output_date( esc_html( $due_date ) ); ?></p>
 	</div>
 
 </div>